So, after looking at this a bit more, I see that the "missing mp4:" is something that JW Player in particular allows:
****
From [ http://www.longtailvideo.com/support/jw-player/jw-player-for-flash-v5/12535/video-delivery-rtmp-streaming ]:

Note that the documentation of RTMP servers tell you to set the file option in players like this:

For FLV video: file=clip (without the .flv extension).
For MP4 video: file=mp4:clip.mp4 (with mp4: prefix).
For MP3 audio: file=mp3:song.mp3 (with mp3: prefix).
For AAC audio: file=mp4:song.aac (with mp4: prefix).

You do not have to do this with the JW Player, since the player takes care of stripping the extension and/or adding the prefix. If you do add the prefix yourself, the player will recognize it and not modify the URL.
****

So, I assume the missing mp4: is something that we could expect to see with other instances of JW Player -- so perhaps it makes sense to deal with it in the generic jwplayer assethandler rather than in a specialized hosthandler section?