jshint fixes
This fixes all but 3 jshint errors in the sherd.js code. These
last three errors are going to be tricky to fix: one of the depends
on the use of eval(), and the other two define callback functions
in a deeply nested loop that depend on data within those loops.