:arrow_up: OpenLayers 2.13.1
This upgrades sherdjs's OpenLayers from 2.0.2 to 2.13.1,
which is the latest in the openlayers 2.x series.

This commit also removes the openlayers source directory,
since I don't think there's much point in keeping that in
here. We do need to include openlayer's css file and images
here, for mediathread to load.