Fix clipform bugs with youtube player
This took me forever to fix, including some wrong turns. The problem
was that, when navigating to new youtube videos, the clipform controls
weren't connected to the new video, because the old Sherd.Video.YouTube
instance is still present -- a new one isn't created.

Creating a new Sherd.Video.YouTube instance for each video (which is
what seems like should happen) is not worth the large effort that would
take -- it goes against SherdJS's current architecture. Instead, I've
found a way to make sure the state stays correct with promises, and lots
of sanity checks/debugging.